requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
apscheduler
python-json-logger
httpx
python-multipart
pytest
pytest-asyncio

